fix on migration failure

+++ database/migrations/2021_11_04_022829_create_clients_table.php
@@ -14,7 +14,7 @@ class CreateClientsTable extends Migration
     public function up()
     {
         Schema::create('clients', function (Blueprint $table) {
-            $table->bigInteger('id');
+            $table->increments('id');
             $table->string('name');
             $table->string('url');
             $table->timestamps();
